project successfully build with custom DateTime scalar in place
